image gallery with django/staticfiles/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-appconf==1.0.5
django-autoslug==1.9.8
django-crispy-forms==1.13.0
django-imagekit==4.0.2
django-taggit==1.5.1
pilkit==2.0
Pillow==8.3.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
django-cleanup==5.2.0
whitenoise==5.3.0
django-filter==21.1
